requirements.txt
flask
flask-sqlalchemy
Flask-Migrate
flask_wtf
email_validator